% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{as_curves_df}
\alias{as_curves_df}
\title{Extract curves from a CCE result}
\usage{
as_curves_df(x)
}
\arguments{
\item{x}{A \code{cce_vs_result} or \code{cce_soc_result}.}
}
\value{
A data frame.
}
\description{
Extract curves from a CCE result
}
