% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cce.R
\name{as_effects_df}
\alias{as_effects_df}
\title{Extract the effects table}
\usage{
as_effects_df(x)
}
\arguments{
\item{x}{A \code{cce_fit} object.}
}
\value{
A data frame.
}
\description{
Extract the effects table
}
