% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{read_cce_spec}
\alias{read_cce_spec}
\title{Read a CCE specification from YAML}
\usage{
read_cce_spec(path)
}
\arguments{
\item{path}{YAML file created by \code{\link[=write_cce_spec]{write_cce_spec()}}.}
}
\value{
A \code{cce_spec} object.
}
\description{
Read a CCE specification from YAML
}
