% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{write_cce_results}
\alias{write_cce_results}
\title{Write CCE result files to disk}
\usage{
write_cce_results(x, path = "outputs")
}
\arguments{
\item{x}{A \code{cce_vs_result} or \code{cce_soc_result}.}

\item{path}{Output directory.}
}
\value{
Invisibly returns \code{path}.
}
\description{
Write CCE result files to disk
}
