% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{write_cce_spec}
\alias{write_cce_spec}
\title{Write a CCE specification to YAML}
\usage{
write_cce_spec(spec, path)
}
\arguments{
\item{spec}{A \code{cce_spec} object.}

\item{path}{File path ending in \code{.yml} or \code{.yaml}.}
}
\value{
Invisibly returns \code{path}.
}
\description{
Write a CCE specification to YAML
}
