% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cce.R
\name{as_diagnostics_df}
\alias{as_diagnostics_df}
\title{Extract the diagnostics table}
\usage{
as_diagnostics_df(x)
}
\arguments{
\item{x}{A \code{cce_fit} object.}
}
\value{
A data frame.
}
\description{
Extract the diagnostics table
}
