% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{as_effects_df}
\alias{as_effects_df}
\title{Extract effect summaries from a CCE result}
\usage{
as_effects_df(x)
}
\arguments{
\item{x}{A \code{cce_vs_result} or \code{cce_soc_result}.}
}
\value{
A data frame.
}
\description{
Extract effect summaries from a CCE result
}
