% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cce.R
\docType{package}
\name{cce-package}
\alias{cce}
\alias{cce-package}
\title{cce: Minimal Counterfactual Comparator Workflows}
\description{
\code{cce} provides a compact set of survival workflows for observed
two-arm comparisons and SOC-only projection scenarios.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dai540/cce}
  \item \url{https://dai540.github.io/cce/}
  \item Report bugs at \url{https://github.com/dai540/cce/issues}
}

}
\author{
\strong{Maintainer}: Dai

}
\keyword{internal}
