% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cce.R
\name{write_cce_results}
\alias{write_cce_results}
\title{Write CCE results to disk}
\usage{
write_cce_results(x, dir)
}
\arguments{
\item{x}{A \code{cce_fit} object.}

\item{dir}{Output directory.}
}
\value{
Invisibly returns \code{dir}.
}
\description{
Write CCE results to disk
}
